pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scientific-rag"
version = "0.1.0"
description = "Hybrid RAG pipeline for scientific and technical document retrieval"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "qdrant-client>=1.9",
    "fastembed>=0.3",
    "sentence-transformers>=2.2",
    "langchain-text-splitters>=0.2",
    "loguru>=0.7",
    "python-dotenv>=1.0",
    "docling>=1.0",
    "boto3>=1.34",
    "tqdm>=4.60",
    "nbformat>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "jupyter>=1.0",
    "ipykernel>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

